[MIG] l10n_es_account_invoice_sequence: Modulo migrado a la v8
