+++ l10n_es_account_invoice_sequence/__openerp__.py
