+++ l10n_es_account_invoice_sequence/i18n/l10n_es_account_invoice_sequence.pot
